Added constructors for test coverage
