+++ core/src/main/java/io/parsingdata/metal/expression/derived/GUID.java
@@ -40,7 +40,7 @@ import io.parsingdata.metal.expression.value.ValueExpression;
 public final class GUID {
     private static final Encoding BIG_ENDIAN = new Encoding();
 
-    private GUID() {}
+    GUID() {}
 
     /**
      * Use a String representation of a GUID as predicate.
